+++ MemfaultPackages/bort/src/main/java/com/memfault/bort/settings/DynamicSettingsProvider.kt
@@ -93,6 +93,8 @@ open class DynamicSettingsProvider(
             get() = settings.bugReportFirstBugReportDelayAfterBoot.duration
         override val rateLimitingSettings: RateLimitingSettings
             get() = settings.bugReportRequestRateLimitingSettings
+        override val periodicRateLimitingPercentOfPeriod: Int
+            get() = settings.bugReportPeriodicRateLimitingPercentOfPeriod
     }
 
     override val dropBoxSettings = object : DropBoxSettings {